refactor(ast): extract operation params into operationParams.ts (#3572 1e)
Move createOperationParams, resolveParamType, buildGroupParam, buildTypeLiteral,
caseParams, and createDiscriminantNode (plus the OperationParamsResolver,
CreateOperationParamsOptions, ParamGroupType, and BuildGroupArgs types and the
caseParamsMemo/buildGroupProperty internals) out of utils/ast.ts into
utils/operationParams.ts, with their tests in operationParams.test.ts. refs.ts,
factory.ts, and types.ts now reference operationParams.ts. utils/ast.ts holds
only the schema-graph helpers; the four @kubb/ast barrels are unchanged.
